requirements.txt
Pyrogram==2.0.106
TgCrypto==1.2.5
requests==2.32.3
python-dotenv==1.0.1